+++ app/components/Settings/Metrics/styledComponents.js
@@ -37,9 +37,26 @@ export const Divider = styled.div`
   width: 100%;
 `;
 
+export const Language = styled.div`
+  align-items: center;
+  display: flex;
+  flex-flow: wrap;
+`;
+
+export const LanguageListItem = styled.div`
+  display: flex;
+  font-size: ${defaultFontSize};
+  padding: 0.5rem 0;
+  justify-content: space-between;
+`;
+
 export const RankingContainer = styled.div`
   display: flex;
   flex-direction: column;
+`;
+
+export const StyledLanguageAutocomplete = styled.div`
+  width: 80%;
 `;
 
 export const StyledUserBarTitle = styled.div`
