+++ app/components/Settings/styledComponents.js
@@ -1,53 +1,25 @@
 import React from 'react';
 import styled from 'styled-components';
-import ExpansionPanel from '@material-ui/core/ExpansionPanel';
-import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
 import Paper from '@material-ui/core/Paper';
 import Popper from '@material-ui/core/Popper';
 import Tab from '@material-ui/core/Tab';
 import Tabs from '@material-ui/core/Tabs';
 
-import {
-  BaseDropDownMenu,
-  BaseTextInput,
-  ErrorSuccessBanner,
-} from 'components/base_ui';
+import { BaseDropDownMenu, ErrorSuccessBanner } from 'components/base_ui';
 import {
   commentHeaderColor,
   defaultFontSize,
   hoverLinkColor,
-  lightBlueColor,
   subTextLightGrey,
   textColor,
 } from 'defaultStyleHelper';
 import { mediaQueriesByDevice } from 'utils/breakpoints';
 
-const { mobile, laptop } = mediaQueriesByDevice;
+const { mobile } = mediaQueriesByDevice;
 
 export const BaseInputWrapper = styled.div`
   margin-left: ${({ hasMargin }) => (hasMargin ? '4rem' : '0')};
   width: 45%;
-`;
-
-export const DetailContainer = styled.div`
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-`;
-
-export const DetailViewContainer = styled.div`
-  color: ${textColor};
-  display: flex;
-
-  ${mobile} {
-    flex-direction: column;
-  }
-`;
-
-export const EditUserImageWrapper = styled.div`
-  align-self: center;
-  position: relative;
 `;
 
 export const EmptyComponentContainer = styled.div`
@@ -75,45 +47,6 @@ export const IconButtonGroup = styled.div`
   display: flex;
 `;
 
-export const InputIconGroup = styled.label`
-  align-items: center;
-  background-color: white;
-  border-radius: 2rem;
-  border: 0.2rem solid ${lightBlueColor};
-  bottom: -1rem;
-  color: ${lightBlueColor};
-  display: flex;
-  height: 4rem;
-  justify-content: center;
-  padding: 1rem 0;
-  position: absolute;
-  right: -1rem;
-  width: auto;
-
-  &:hover {
-    background-color: white;
-    cursor: pointer;
-  }
-
-  svg {
-    height: 2rem;
-    width: 2rem;
-  }
-`;
-
-export const Language = styled.div`
-  align-items: center;
-  display: flex;
-  flex-flow: wrap;
-`;
-
-export const LanguageListItem = styled.div`
-  display: flex;
-  font-size: ${defaultFontSize};
-  padding: 0.5rem 0;
-  justify-content: space-between;
-`;
-
 export const LinkIcon = styled.div`
   padding-right: 0.5rem;
 
@@ -130,37 +63,17 @@ export const LinksWrapper = styled.div`
   margin-left: 0.5rem;
 `;
 
-export const Name = styled.div`
-  font-size: 2rem;
-  font-weight: 500;
-  line-height: 1.2;
-  margin-bottom: 0.5rem;
-  margin-left: 0.5rem;
-`;
-
-export const OneLink = styled.div`
-  align-items: center;
-  display: flex;
-`;
-
-export const OneLinkWrapper = styled.div`
-  align-items: center;
-  display: flex;
-  padding: 0.5rem 0;
-  justify-content: space-between;
-`;
-
-export const Rep = styled.div`
-  align-items: center;
-  display: flex;
-  font-size: 1.6rem;
-  margin: 0.5rem 0 0.5rem 0.5rem;
-`;
-
 export const SearchContainer = styled.div`
   display: flex;
   justify-content: space-between;
   margin-bottom: 3rem;
+  width: 100%;
+`;
+
+export const SettingsContainer = styled.div`
+  align-items: center;
+  display: flex;
+  flex-direction: column;
   width: 100%;
 `;
 
@@ -175,55 +88,17 @@ export const SettingsTabsWrapper = styled.div`
   }
 `;
 
-export const StyledA = styled.a`
-  &:hover {
-    text-decoration: underline;
-  }
+export const SettingsViewContainer = styled.div`
+  background-color: white;
+  padding: 4rem 2rem 2rem;
 `;
 
 export const StyledBaseDropDownMenu = styled(BaseDropDownMenu)`
   margin-left: 1rem;
 `;
 
-export const StyledBaseTextInput = styled(BaseTextInput)`
-  margin: 0;
-`;
-
 export const StyledErrorSuccessBanner = styled(ErrorSuccessBanner)`
   margin-top: 1.6rem;
-  width: 100%;
-`;
-
-export const StyledExpansionPanel = styled(ExpansionPanel)`
-  box-shadow: none;
-  color: ${textColor};
-
-  &:before {
-    background-color: white;
-  }
-
-  &.expanded {
-    margin: 0rem;
-  }
-
-  .MuiExpansionPanelSummary-root {
-    margin: 0;
-    padding: 0;
-  }
-
-  .MuiExpansionPanelSummary-content {
-    margin: 0;
-  }
-
-  svg {
-    height: 2rem;
-    width: 2rem;
-  }
-`;
-
-export const StyledExpansionPanelDetails = styled(ExpansionPanelDetails)`
-  justify-content: center;
-  padding: 0;
   width: 100%;
 `;
 
@@ -232,10 +107,6 @@ export const StyledH3 = styled.h3`
   font-size: 2rem;
   font-weight: 500;
   margin: ${({ removeMarginTop }) => (removeMarginTop ? '0 0 3rem' : '3rem 0')};
-`;
-
-export const StyledLanguageAutocomplete = styled.div`
-  width: 80%;
 `;
 
 export const StyledPaper = styled(Paper)`
@@ -273,15 +144,13 @@ export const StyledTab = styled(Tab)`
 export const StyledTabs = styled(({ displayBottom, ...restProps }) => (
   <Tabs {...restProps} />
 ))`
-  margin: 4rem 0 0;
-
   ${mobile} {
     display: ${({ displayBottom }) => (displayBottom ? 'none' : 'flex')};
     margin-top: 2rem;
   }
 
   .indicator {
-    background-color: transparent;
+    display: none;
   }
 
   svg {
@@ -304,36 +173,4 @@ export const TabItemBorder = styled.div`
     cursor: pointer;
     border-bottom: 0.2rem solid ${commentHeaderColor};
   }
-`;
-
-export const UserCardWrapper = styled.div`
-  background-color: white;
-  border-radius: 0.2rem;
-  display: flex;
-  flex-direction: column;
-  height: fit-content;
-  padding-right: 5rem;
-  width: fit-content;
-
-  ${laptop} {
-    display: ${({ displayBottom }) => (displayBottom ? 'none' : 'flex')};
-    padding-right: 0;
-    width: 100%;
-  }
-`;
-
-export const UserImage = styled.img`
-  height: 25rem;
-  margin-bottom: 0.5rem;
-  object-fit: cover;
-  width: 25rem;
-
-  ${mobile} {
-    align-self: center;
-  }
-`;
-
-export const ComponentContainer = styled.div`
-  background-color: white;
-  padding: 4rem 2rem 2rem;
 `;
